Added CSV parsing.
* Imported Apache Commons CSV library into build.gradle.
* Added CSV parsing logic into CsvMunroDataLoader.
  * Added basic unit tests for this.
* Updated GridReference class to use factory method.
* Fixed order-of-operations in REGEX assertion in GridReference constructor.
* Changed height in Munro class to double.
* Added display names for Designation class to aid readability.

+++ src/main/java/com/com/munroes/model/Munro.java
@@ -8,6 +8,6 @@ public final class Munro {
 
     private final @NonNull String name;
     private final @NonNull GridReference gridReference;
-    private final int height;
+    private final double height;
     private final @NonNull Designation designation;
 }
